requirements.txt
streamlit
pandas
yahooquery
openai
python-dotenv
langchain
langchain-community
langchain_mistralai
